Fix bug in finding next blank row.
Originally there were no formulas in the sheet data area, so we could find the blank row by asking for the used range.  But the unallocated pledge formulas now make that impossible (and point out the failure that happens if people insert randowm data).

So now we count blanks in the data area of the first column, and then subtract that from the end of the used range.  This introduces a few possible edge cases:

1. There's a blank row in the middle of filled data area.  This will cause the last row to be overwritten.
2. The first row of the data area doesn't match the settings value.  This may lead to an incorrect count.
3. The unallocated pledge formula doesn't extend to the end of the sheet.  This will cause the data area and the filled are to go out of sync.

All of these seem like mininmal risks.